requirements.txt
# FastAPI and web server
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
pydantic>=2.11.7
pydantic-settings>=2.10.1

# Google Cloud Pub/Sub
google-cloud-pubsub>=2.31.1
google-auth>=2.40.3
google-api-core>=2.25.1

# Additional utilities
python-dotenv>=1.1.1
structlog>=25.4.0
httpx>=0.28.1

# Development and testing (optional)
pytest>=8.4.1
pytest-asyncio>=1.1.0
black>=25.1.0
isort>=6.0.1